Move ecs::flags unittest into its source file
